requirements.txt
nasbench_pytorch
ConfigSpace
simplejson
yacs
pathvalidate

torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
